🚀 chore(package.json): add react-icons and react-paginate dependencies 🔧 fix(poster.tsx): make the link prop optional in PosterProps interface 🔧 fix(WatchList.tsx): update styling and add pagination functionality 🔧 fix(watchListSlice.tsx): add currentPage and moviesPerPage to initial state 🔧 fix(AddWatchBtn.tsx): change isThisInWatchList type to string in ButtonProps
The changes were made to enhance the functionality and user experience of the application. Dependencies were added for additional features, the link prop was made optional for flexibility, styling and pagination were improved in the WatchList page, and the initial state in watchListSlice was updated to include currentPage and moviesPerPage for better state management. Additionally, the type of isThisInWatchList was changed to string in ButtonProps for consistency.