🔧 fix(RatePopup.tsx): change ModalBox and ModalContainer to dialog and form 🚀 feat(RatePopup.tsx): add X icon from lucide-react and support for process.env.PORT environment variable
The commit changes the styled components ModalBox and ModalContainer to use the HTML dialog and form elements for better semantic structure. Additionally, an X icon from lucide-react is added for the close button in the RatePopup component. The support for the process.env.PORT environment variable is added to make the application more flexible in terms of port configuration.

🔧 chore(Header.tsx): update lucide-react icons in Header component
🔧 chore(SearchBar.tsx): remove commented out code in SearchBar component
The lucide-react icons in the Header component have been updated to use the Clapperboard icon instead of the Video icon for the IMDb link. This change improves consistency and visual appeal. Additionally, unnecessary commented out code has been removed from the SearchBar component for better code cleanliness.